requirements.txt
python-dotenv==1.0.1
boto3==1.35.7
requests==2.32.3
textblob==0.18.0.post0
pandas==2.2.2